requirements.txt
requests==2.25.1
pymongo==3.11.3
python-crontab==2.5.1